Introduce ExecutionCancelled exception and split test files
- Add ExecutionCancelled exception for get_result() on cancelled tasks,
  avoiding conflation with asyncio.CancelledError semantics
- Split test_dependencies_core.py into focused test files:
  - test_dependency_uniqueness.py (single=True constraint tests)
  - test_handler_semantics.py (FailureHandler/CompletionHandler behavior)
  - test_dependencies_core.py (core dependency injection tests)
